requirements.txt
matplotlib==3.7.2
numpy==1.24.3
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.24.1
tensorflow==2.13.0
tensorflow_macos==2.13.0
